Reporte/AZURE/2transform/requirements.txt
azure-functions
azure-storage-blob
pandas
pyarrow
